tasks now deletable

+++ LB2/todo-list-node/user/tasklist.js
@@ -16,7 +16,6 @@ async function getHtml(req) {
     let conn = await db.connectDB();
     const sql = 'select ID, title, state from tasks where UserID = ?'
     let [result, fields] = await conn.query(sql, req.cookies.userid);
-
     result.forEach(function(row) {
         html += `
             <tr>
@@ -24,7 +23,7 @@ async function getHtml(req) {
                 <td class="wide">`+row.title+`</td>
                 <td>`+ucfirst(row.state)+`</td>
                 <td>
-                    <a href="edit?id=`+row.ID+`">edit</a> | <a href="delete?id=`+row.ID+`">delete</a>
+                    <a href="edit?id=`+row.ID+`">edit</a> | <a href="delete?id=`+row.ID+`" onclick="return confirm('Are you sure you want to delete this item?');">delete</a>
                 </td>
             </tr>`;
     });
